refactor: update index.js file for initial setup
This commit adds the initial setup for the project. The `index.js` file imports the `createHomePage` function from the `homePage.js` module and defines an `initialLoad` function.
The 'initialLoad' function calls the `createHomePage` function. This is to set up the basic structure to dynamically generate the content of the homepage.

Next step will be to implement the `createHomePage` function inside the `homePage.js` file to generate the actual content.