refactor: index.js to dynamically generate content
Having re-read the instructions of the project, my initial approach was going to be to add an unordered list into the html to append the menu and contact tabs to,
however I now believe this is not the correct approach as to what the project is asking for. In the instructions it states:

Put the contents of each ‘tab’ inside of its own module. Each module will export a function that creates a div element,
adds the appropriate content and styles to that element and then appends it to the DOM.

Hence the following changes:

- Remove unnecessary homeTab and menuTab variables, as they will not be required.
- Modify the initialLoad() function to create the contentDiv element using document.querySelector('#content') and pass it as an argument to createHomePage(contentDiv).

I believe this change better aligns with the project instructions, which emphasises generating content dynamically rather than relying on hardcoded HTML elements.
The switchTab() function remains unchanged, it is still responsible for switching between the "home" and "menu" tabs by clearing the content and calling the respective page creation functions.