add: anchor tag for menu switching functionality
This commit adds the menu tab element and attaches an event listener, the user will be able to click on the "Menu" tab and have the switchTab function called with the argument 'menu',
which will then be populated with the menu page content. This change allows for tab-based navigation and dynamic content loading.