feat: add switchTab argument to menu case
Add switchTab as an argument in the 'menu' case, this allows access to the switchTab function defined in index.js.
Therefore allowing the createMenuPage function to attach the event listener to the "Home" tab on the menu page, this enables the user to switch back to the home page when clicked.