refactor: createTabs function & update export
This commit changes the createTabs function to use the default export syntax, it was previously exported as a named export.
This change improves code readability by moving the export to the bottom of the function which makes it easier to see.
It also helps adhere to the pattern of using default exports for single primary functions.